chore: address issue #608
